feature-add-qr-code-scanner-branch

+++ loyalty_app/lib/Features/home_layout/presentation/views/widgets/offer_widgets/offers_item.dart
@@ -27,7 +27,7 @@ class OffersItem extends StatelessWidget {
   final String? coverImage;
   final String? storeImage;
   final String? offerName;
-  final int? pointsAmount;
+  final double? pointsAmount;
   final String? storeName;
   final GestureTapCallback onTap;
